+++ packages/orq-rc/RELEASES.md
@@ -8899,3 +8899,13 @@ Based on:
 - [typescript v4.9.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.28 - packages/orq-rc
+
+## 2026-05-07 04:18:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.29 - packages/orq-rc